Update Mixer binary directory names. (#2043)
