+++ mixer/cmd/mixc/cmd/check.go
