+++ mixer/cmd/mixcol/main.go
@@ -22,7 +22,7 @@ package main
 import (
 	"os"
 
-	"istio.io/istio/mixer/cmd/collateral/cmd"
+	"istio.io/istio/mixer/cmd/mixcol/cmd"
 	"istio.io/istio/mixer/cmd/shared"
 )
