+++ mixer/cmd/mixs/cmd/inventory.go
