+++ mixer/cmd/mixs/main.go
@@ -17,12 +17,28 @@ package main
 import (
 	"os"
 
-	"istio.io/istio/mixer/cmd/client/cmd"
+	adapter "istio.io/istio/mixer/adapter"
+	"istio.io/istio/mixer/cmd/mixs/cmd"
 	"istio.io/istio/mixer/cmd/shared"
+	adptr "istio.io/istio/mixer/pkg/adapter"
+	"istio.io/istio/mixer/pkg/template"
+	generatedTmplRepo "istio.io/istio/mixer/template"
 )
 
+func supportedTemplates() map[string]template.Info {
+	return generatedTmplRepo.SupportedTmplInfo
+}
+
+func supportedAdapters() []adptr.InfoFn {
+	return adapter.Inventory()
+}
+
+func supportedLegacyAdapters() []adptr.RegisterFn {
+	return adapter.InventoryLegacy()
+}
+
 func main() {
-	rootCmd := cmd.GetRootCmd(os.Args[1:], shared.Printf, shared.Fatalf)
+	rootCmd := cmd.GetRootCmd(os.Args[1:], supportedTemplates(), supportedAdapters(), supportedLegacyAdapters(), shared.Printf, shared.Fatalf)
 
 	if err := rootCmd.Execute(); err != nil {
 		os.Exit(-1)
