+++ mixer/doc/running-local-mixer.md
@@ -10,7 +10,7 @@ The `KUBECONFIG` environment variable specifies the location of the Kubernetes c
 If you do not want to use the Kubernetes adapter, remove `mixer/testdata/configroot/scopes/global/subjects/global/rules.yml`.
 
 ```shell
-KUBECONFIG=${HOME}/.kube/config bazel-bin/mixer/cmd/server/mixs server --logtostderr --configStore2URL=fs://$(pwd)/mixer/testdata/config --configStoreURL=fs://$(pwd)/mixer/testdata/configroot  -v=4
+KUBECONFIG=${HOME}/.kube/config bazel-bin/mixer/cmd/mixs/mixs server --logtostderr --configStore2URL=fs://$(pwd)/mixer/testdata/config --configStoreURL=fs://$(pwd)/mixer/testdata/configroot  -v=4
 ```
 
 You can also run a simple client to interact with the server:
@@ -20,7 +20,7 @@ Note that `source.ip` is an ip address specified as 4 `:` separated bytes.
 `192.0.0.2` is encoded as `c0:0:0:2` in the example.
 
 ```shell
-bazel-bin/mixer/cmd/client/mixc check -v 2 --string_attributes destination.service=abc.ns.svc.cluster.local,source.name=myservice,target.port=8080 --stringmap_attributes "request.headers=clnt:abcd;source:abcd,destination.labels=app:ratings,source.labels=version:v2"   --timestamp_attributes request.time="2017-07-04T00:01:10Z" --bytes_attributes source.ip=c0:0:0:2
+bazel-bin/mixer/cmd/mixc/mixc check -v 2 --string_attributes destination.service=abc.ns.svc.cluster.local,source.name=myservice,target.port=8080 --stringmap_attributes "request.headers=clnt:abcd;source:abcd,destination.labels=app:ratings,source.labels=version:v2"   --timestamp_attributes request.time="2017-07-04T00:01:10Z" --bytes_attributes source.ip=c0:0:0:2
 
 Check RPC completed successfully. Check status was OK
   Valid use count: 10000, valid duration: 5m0s
@@ -28,7 +28,7 @@ Check RPC completed successfully. Check status was OK
 
 The following command sends a `report` request to Mixer.
 ```shell
-bazel-bin/mixer/cmd/client/mixc report -v 2 --string_attributes destination.service=abc.ns.svc.cluster.local,source.name=myservice,target.port=8080 --stringmap_attributes "request.headers=clnt:abc;source:abcd,destination.labels=app:ratings,source.labels=version:v2"  --int64_attributes response.duration=2003,response.size=1024 --timestamp_attributes  request.time="2017-07-04T00:01:10Z" --bytes_attributes source.ip=c0:0:0:2
+bazel-bin/mixer/cmd/mixc/mixc report -v 2 --string_attributes destination.service=abc.ns.svc.cluster.local,source.name=myservice,target.port=8080 --stringmap_attributes "request.headers=clnt:abc;source:abcd,destination.labels=app:ratings,source.labels=version:v2"  --int64_attributes response.duration=2003,response.size=1024 --timestamp_attributes  request.time="2017-07-04T00:01:10Z" --bytes_attributes source.ip=c0:0:0:2
 
 Report RPC returned OK
 ```
