+++ tools/setup_run
@@ -10,7 +10,7 @@ cd istio; set +x; source bin/use_bazel_go.sh ; set -x; cd ..
 fortio server &
 ( cd proxy/src/envoy/mixer; ./start_envoy > /tmp/envoy.log ) &
 # add -v=5 for verbose/debug
-./istio/bazel-bin/mixer/cmd/server/mixs server --configStore2URL=fs://$(pwd)/mixerconfig --configStoreURL=fs://$(pwd)/emptydir --logtostderr 2> /tmp/mixs.2.log > /tmp/mixs.1.log &
+./istio/bazel-bin/mixer/cmd/mixs/mixs server --configStore2URL=fs://$(pwd)/mixerconfig --configStoreURL=fs://$(pwd)/emptydir --logtostderr 2> /tmp/mixs.2.log > /tmp/mixs.1.log &
 echo "starting everything..."
 sleep 3
 curl -v http://localhost:9090/debug
